backend/pyproject.toml
[project]
name = "datapipeline-backend"
version = "0.1.0"
description = "Unified Data Pipeline Platform — FastAPI backend"
requires-python = ">=3.12"
readme = "../README.md"

dependencies = [
  # HTTP framework
  "fastapi>=0.115,<0.116",
  "uvicorn[standard]>=0.32,<0.33",

  # Settings / validation
  "pydantic>=2.9,<3",
  "pydantic-settings>=2.6,<3",

  # HTTP client (async)
  "httpx>=0.27,<0.29",

  # Security
  "argon2-cffi>=23.1,<25",
  "python-jose[cryptography]>=3.3,<4",

  # Logging
  "structlog>=24.4,<25",

  # Observability
  "prometheus-client>=0.21,<0.22",

  # NOTE: DB 드라이버(sqlalchemy[asyncio], asyncpg, alembic)는
  # Phase 1.2.3 (DB & Migration) 에서 추가한다.
]

[dependency-groups]
dev = [
  "ruff>=0.7,<0.9",
  "mypy>=1.11,<1.13",
  "pytest>=8.3,<9",
  "pytest-asyncio>=0.24,<0.25",
  "pytest-cov>=5.0,<6",
  "httpx>=0.27,<0.29",
  "types-python-jose>=3.3,<4",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
  "E", "F", "W",     # pycodestyle / pyflakes
  "I",               # isort
  "UP",              # pyupgrade
  "B",               # bugbear
  "ASYNC",           # async lint
  "SIM",             # simplify
  "RUF",             # ruff specific
]
ignore = [
  "E501",            # line length (line-length 이미 적용)
  "B008",            # FastAPI Depends() 허용
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
disallow_untyped_decorators = false   # FastAPI route decorator용
warn_unused_ignores = true
show_error_codes = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["jose.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --strict-config"
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
source = ["app"]
branch = true

[tool.coverage.report]
show_missing = true
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
